+++ scripts/dialogs/export.mjs
@@ -35,7 +35,7 @@ export class ExportDialog extends FormApplication {
 	static get defaultOptions() {
 		return {
 			...super.defaultOptions,
-			title: `${MODULE.TITLE} - Export Modules`,
+			title: `${MODULE.TITLE} - ${MODULE.localize('dialog.titles.export')}`,
 			id: `${MODULE.ID}-export-dialog`,
 			classes: ['dialog'],
 			template: `./modules/${MODULE.ID}/templates/export.hbs`,
@@ -53,9 +53,9 @@ export class ExportDialog extends FormApplication {
 			}, 
 			activeModules: () => {
 				// Build Markdown Display
-				let markdown = ['### Active Modules'];
+				let markdown = [`### ${MODULE.localize('dialog.activeModules')}`];
 				for (const [key, value] of Object.entries(this.moduleData)) {
-					markdown.push(`${value.title} v${value.version}` + (value.bugs ? ' [Bugs URL]('+value.bugs+')' : ''));
+					markdown.push(`${value.title} v${value.version}` + (value.bugs ? ` [${MODULE.localize('dialog.bugsUrl')}](${value.bugs})` : ''));
 				}
 				return markdown.join('\n');
 			}
@@ -68,7 +68,7 @@ export class ExportDialog extends FormApplication {
 		$(html).find('[data-button="copy"]').on('click', (event) => {
 			$(html).find(`#${MODULE.ID}-copy-export`).select();
 			document.execCommand('copy');
-			ui.notifications.info(`<strong>${MODULE.TITLE}</strong>: Active Module List has been Copied to Your Clipboard!`);
+			ui.notifications.info(`<strong>${MODULE.TITLE}</strong>: ${MODULE.localize('activeModulesCopied')}`);
 		})
 
 		$(html).find('[data-button="export"]').on('click', (event) => {